+++ src/lib.rs
@@ -1,5 +1,3 @@
-mod ack_channel;
-mod dag;
 mod path;
 mod planner;
 mod system;
@@ -12,4 +10,4 @@ pub mod workflow;
 // TODO: this should not be exported from this crate.
 // It would more sense to re-export it, including the seq
 // and dag macros, from a "gustav-test" crate
-pub use dag::Dag;
+pub use workflow::Dag;
